* Vulkan API modernization (part #3):    * vulkan_compute/vulkan_renderer: redesign/improve/fix the whole swapchain and present handling:      * vulkan_compute: VK_KHR_get_surface_capabilities2 and VK_EXT_surface_maintenance1 are now required instance extensions (if !console-only)      * vulkan_compute: VK_EXT_swapchain_maintenance1 is now a required device extension (if !console-only)      * vulkan_compute: this now creates semaphores/fences for swapchain image acquisition and present handling and makes these available externally (+uses them internally)      * vulkan_compute: acquire_next_image() now deals with all of the sema/fence setup (+fixed the potential race)      * vulkan_compute: acquire_next_image() now longer performs an image layout transition, this now needs to be done by an external caller (we do track the current/previous layout via a new swapchain_prev_layouts vector now) -> this is no longer a blocking call as long as this doesn't have to wait on previous swapchain images and/or fences to become available      * vulkan_compute: queue_present() now waits on a present semaphore (if available) and signals a present fence      * vulkan_renderer: when a swapchain image (drawable) is specified (via a previous get_next_drawable() call), the renderer will now perform the swapchain image layout transition at the very start      * vulkan_renderer: make use of new acquisition/present semaphores (wait on swapchain image acquisition, signal present semaphore after renderer queue submit has finished)      * vulkan_renderer: ignore cur_drawable in set_attachment(), this is handled separately now      * vulkan_compute: make use of new VK_EXT_swapchain_maintenance1 functionality so that we can use a present_fence to know when a present has actually finished      * NOTE: this fixes the remaining validation issues that were still present + should be more async/faster now    * vulkan_compute: moved VR specific acquisition from acquire_next_image() into a separate acquire_next_vr_image() function    * vulkan_compute/vulkan_device: we now determine and store the Vulkan physical device index (inside the parent context/instance)    * vulkan_compute: added vulkan_insert_cmd_debug_label() support -> vkCmdInsertDebugUtilsLabelEXT    * vulkan_queue: avoid using vulkan_insert_cmd_debug_label()/vulkan_end_cmd_debug_label() for execute_indirect() / secondary command buffer usage, this seems to trigger an issue in validation layers -> use new vulkan_insert_cmd_debug_label() instead    * vulkan_queue/vulkan_kernel: place vulkan_end_cmd_debug_label() at more appropriate locations    * vulkan_encoder: the debug label is now stored inside the vulkan_encoder object    * compute_fence: added Vulkan specific TOP_OF_PIPE and BOTTOM_OF_PIPE sync stages    * vulkan_common: added TODO that the v2 usage flags bug has been fixed now and the workaround should be removed with the next Vulkan SDK release    * vulkan_image: use/check the "is_aliasing"/VULKAN_ALIASING flag rather than the "is_aliased_array" flag to determine whether a dedicated allocation can be used  * metal_renderer: for compat, BOTTOM_OF_PIPE is mapped to VERTEX and TOP_OF_PIPE to FRAGMENT  * cuda_image: fixed log command call

+++ compute/cuda/cuda_image.cpp
@@ -322,7 +322,7 @@ bool cuda_image::create_internal(const bool copy_host_data, const compute_queue&
 		// copy host memory to device if it is non-null and NO_INITIAL_COPY is not specified
 		if(copy_host_data && host_data.data() != nullptr && !has_flag<COMPUTE_MEMORY_FLAG::NO_INITIAL_COPY>(flags)) {
 			log_debug("copying $ bytes from $X to array $X",
-					  image_data_size, host_data.data(), image);
+					  image_data_size, uint64_t(host_data.data()), image);
 			auto cpy_host_data = host_data;
 			apply_on_levels([this, &cpy_host_data](const uint32_t& level,
 												   const uint4& mip_image_dim,
